Main functionality complete
The main functionality works, but the output file also reverses the way the lines are being read.
example
input file:
123
345
435
output file: -> should be
534                  321
543                  543
321                  543

also need to create main driver. this currently only works with ints. Create a driver that will create a separate type stack when reading in the line?